requirements.txt
# NoneBot2 核心
nonebot2[fastapi]>=2.4.3
nonebot-adapter-onebot>=2.4.0

# NoneBot 插件
nonebot-plugin-apscheduler>=0.4.0
nonebot-plugin-status>=0.9.0

# HTTP 客户端
httpx>=0.27.0

# AI 生成
google-generativeai>=0.3.0

# 其他 NoneBot 插件或功能依赖
pillow>=10.2.0
requests>=2.31.0
qrcode>=7.4.2
aiofiles>=23.2.1
python-dotenv>=1.0.0
faster-whisper>=1.0.3

# 时区支持
pytz>=2024.1
tzdata>=2024.1

# Web 管理界面依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 其他依赖
pydantic>=2.0.0
setuptools>=65.0.0
psutil>=5.9.0  # nonebot-plugin-status 需要
